requirements.txt
django 
djangorestframework
httpie
gunicorn
